Create Valid Account Number.java
Using a text editor, create a file (account.txt) that contains a list of at
least 10 six-digit account numbers. Read in each account number and
display whether it is valid. An account number is valid only if the last
digit is equal to the sum of the first five digits divided by 10.
For example,
the number 223355 is valid because the sum of the first five digits
is 15, the remainder when 15 is divided by 10 is 5, and the last digit
is 5.
Write only valid account numbers to an output file (valid.txt).