Create Threads Example 2
Write a Java program that creates two threads T1 and T2 by
extending Thread class. T1 asks the user to enter an integer and
displays a statement that indicates whether the integer is even or odd.
T2 asks the user to enter three integers and displays them in ascending
and descending order.